pyproject.toml
[project]
name = "aegis-test-agent"
version = "0.1.0"
description = "Aegis Test Agent Project"
authors = [
    {name = "Pedro Visconti Guidotte",email = "peguidotte@email.com"}
]
requires-python = "^3.13"
dependencies = [
    "google-cloud-pubsub (>=2.34.0,<3.0.0)",
    "aio-pika (>=9.5.8,<10.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "aegis_agents", from = "src"}]

[tool.pyright]
reportMissingImports = false
reportMissingTypeStubs = false
